+++ apps/client/src/features/app-settings/panel/client-control-panel/ClientControlPanel.module.scss
@@ -13,10 +13,3 @@
     margin-right: $element-spacing;
   }
 }
-
-.buttonSection {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: $element-spacing;
-  margin-top: -$element-spacing;
-}
